Stun enemies when they are hit
Added a stun variable to the slash script for each of the charge states
and to the enemy script. The enemy's stun variable is set when it
collides with a slash, and it becomes unable to move until it reaches 0.

Also added a drag variable to the enemy script, that subtracts a
percentage of the velocity even when the enemy is stunned, so it can
still deacelerate.