Make player collide with walls
Changed the collision shape of the walls and player, the collision masks
for the slash, enemy, and walls, made the player a KinematicBody2D
instead of an Area2D, and made the necessary changes in the player's
script to make the player (and enemies) collide with walls properly.

Also reorganized the player's script for consistency, and removed the
unused "ScreenSize" vector.