Make GameManager a separate node from Main node
Added a new node to the "Main" scene that has the GameManager script,
and removed the script from the root node of the scene. This fixes a bug
where the "Main" scene couldn't be reloaded.
Also made the "Initialize" function of the GameManager script be called
every time the node enters a tree, so that the player variable will be
reassigned once the scene is reloaded and the player node is deleted.