Redo movement code
Removed the old movement code for both the enemies and player along with
all the now unnecessary variables related to it.
Made an interface called "IInput", that has a direction vector, and made
new scripts called "PlayerInput" and "EnemyInput" that implement it.
Made a "Movable" script, that the "Player" and "Enemy" scripts inherit,
that has the new code for movement, using the direction from the IInput
implementation in the path "_inputPath".