Make enemy not move towards player when stunned
After remaking the movement system, the stun no longer worked as it was
part of the old move function. Now the stun is decreased in the process
function, and a new "SetTarget" function has been created to switch the
target to null when the enemy is stunned, and make it target the player
again when it's not stunned and the player instance is valid.