Store highscore and highest wave locally
Made the GameManager store the wave and score variables to a .csv file
in the user's machine when the player dies (only if they're higher than
the value present in the file, if it already exists), to keep track of
the highest score and wave number reached. Also added two new labels to
the title screen to show these values.