Fix player getting knocked back too far
Fixed a bug where the player would be knocked back very far after
hitting a spider behind them with a fully charged slash, and then would
be stuck sliding on the walls for a few seconds. The magnitude of the
player's velocity is now limited to the max speed after hitting a wall,
and the knockback from the enemy uses the normalized velocity rather
than the normal velocity (so the enemy velocity doesn't affect the force
of the knockback, only the direction).