Add shader to see spiders behind objects
Made a shader that displays a checkerboard pattern when the sprite below
is covered by some other sprite. Added the shader to the spiders, so
that they can be seen when they go behind other objects. Edited the
NavigationPolygon, so that the spiders can walk behind objects, now that
they will remain visible when doing so.