Increase enemy hitbox size for player damage
Added a new Area2D to the small spider to be used instead of the
KinematicBody2D's collider to damage the player, and increased it's
size. Changed the signal that the Player script's OnHit function is
connected to, so that it checks for the hitbox's area_entered rather
than body_entered, so that the new Area2D does damage rather than the
enemy's KinematicBody2D collider.

This was done so that the spider's collider for colliding with walls is
separate from the one that damages the player, making it possible for
the spiders to hit player even if they push against a wall vertically
(which would previously make most of the player's collider be inside the
wall, and the spiders couldn't touch it).