Make spiders target the player's foot
Created a new Node2D in the Player scene that represents the foot
position, and made the spider target that rather than the center of the
player.

This was done because, if the player pushed themselves against a wall
vertically near a corner, the spiders would try to reach the player from
the side rather than from below, and that would result in them not being
able to reach the player. Now the spiders will correctly approach from
below in this situation, so the player can no longer remain in these
spots indefinetely.